+++ CHANGELOG.md
@@ -80,6 +80,7 @@
 - Dev: Fixed benchmarks segfaulting on run. (#5559)
 - Dev: Refactored `MessageBuilder` to be a single class. (#5548)
 - Dev: Recent changes are now shown in the nightly release description. (#5553, #5554)
+- Dev: The timer for `StreamerMode` is now destroyed on the correct thread. (#5571)
 
 ## 2.5.1